my_sprinkler_project/raspberry_pi/requirements.txt
paho-mqtt>=1.6.1
google-generativeai>=0.5.0
Adafruit-ADS1x15>=1.0.2
adafruit-circuitpython-dht>=3.7.0
smbus2>=0.4.2
RPi.GPIO>=0.7.1
picamera2>=0.3.12
Pillow>=10.0.0
python-dotenv>=1.0.0
pyserial>=3.5
adafruit-circuitpython-ssd1306>=2.12.0
adafruit-blinka>=8.0.0
